move the login business logic to mixins so that overrides of html and styles are easier
